refactoring by katana on IAM module to rise decoupling level
