+++ gvengine/gviam-db/pom.xml
@@ -9,10 +9,10 @@
     <version>4.0.0-SNAPSHOT</version>
   </parent>
 
-    <artifactId>gviam</artifactId>
+    <artifactId>gviam-db</artifactId>
     <packaging>bundle</packaging>
 
-    <name>GreenVulcano IAM</name>
+    <name>GreenVulcano IAM - DB Implementation</name>
     
 	<dependencies>
 		<dependency>
@@ -27,7 +27,17 @@
 			<version>${version.karaf}</version>
 			<scope>provided</scope>
 		</dependency>
-		
+		<dependency>
+			<groupId>org.hibernate</groupId>
+			<artifactId>hibernate-core</artifactId>
+			<version>5.2.8.Final</version>
+			<scope>provided</scope>
+		</dependency>
+		<dependency>
+			<groupId>it.greenvulcano.gvesb</groupId>
+			<artifactId>gviam</artifactId>
+			<version>${project.version}</version>
+		</dependency>
 	</dependencies>
 
     <build>
@@ -41,18 +51,14 @@
                     <instructions>
                         <Bundle-SymbolicName>${project.artifactId}</Bundle-SymbolicName>
                         <Bundle-Version>${project.version}</Bundle-Version>
-                        <Export-Package>
-                        		it.greenvulcano.gvesb.iam.domain,
-                        		it.greenvulcano.gvesb.iam.jaas,
-                        		it.greenvulcano.gvesb.iam.service,
-                        		it.greenvulcano.gvesb.iam.modules,	
-                        		it.greenvulcano.gvesb.iam.exception
-                        </Export-Package>
-                         <Private-Package>
-                         		it.greenvulcano.gvesb.iam.modules.internal
-                         </Private-Package>
-                                        
-                        <Import-Package>org.apache.karaf.jaas.config,*</Import-Package>
+                        <Export-Package />                        	
+                        <Private-Package>
+                                it.greenvulcano.gvesb.iam.domain.jpa,                        		
+                        		it.greenvulcano.gvesb.iam.repository,
+                        		it.greenvulcano.gvesb.iam.repository.hibernate,
+                        		it.greenvulcano.gvesb.iam.service.internal,
+                        </Private-Package>                        
+                        <Import-Package>org.hibernate.proxy,javassist.util.proxy,org.apache.karaf.jaas.config,*</Import-Package>
                     </instructions>
                 </configuration>
             </plugin>
