+++ gvengine/gviam-db/src/main/java/it/greenvulcano/gvesb/iam/domain/jpa/RoleJPA.java
@@ -1,0 +1,123 @@
+/*******************************************************************************
+ * Copyright (c) 2009, 2016 GreenVulcano ESB Open Source Project.
+ * All rights reserved.
+ *
+ * This file is part of GreenVulcano ESB.
+ *
+ * GreenVulcano ESB is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Lesser General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * GreenVulcano ESB is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ * GNU Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public License
+ * along with GreenVulcano ESB. If not, see <http://www.gnu.org/licenses/>.
+ *******************************************************************************/
+package it.greenvulcano.gvesb.iam.domain.jpa;
+
+import java.io.Serializable;
+
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.Table;
+
+@Entity
+@Table(name="roles")
+public class RoleJPA extends it.greenvulcano.gvesb.iam.domain.Role implements Serializable {
+	private static final long serialVersionUID = 1L;
+	
+	@Id @GeneratedValue(strategy=GenerationType.AUTO)
+	private Integer id;
+	
+	@Column(nullable=false,length=32, unique=true, updatable=false)
+	private String name;
+	
+	@Column(length=320)
+	private String description;
+	
+	public RoleJPA(){
+		
+	}
+	
+	public RoleJPA(String name, String description) {
+		this.name = name;
+		this.description = description;
+	}		
+	
+	@Override
+	public Integer getId() {
+		return id;
+	}
+
+	public void setId(Integer id) {
+		this.id = id;
+	}
+	
+	@Override
+	public String getName() {
+		return name;
+	}
+
+	public void setName(String name) {
+		this.name = name;
+	}
+	
+	@Override
+	public String getDescription() {
+		return description;
+	}
+
+	public void setDescription(String description) {
+		this.description = description;
+	}	
+
+	@Override
+	public int hashCode() {
+		final int prime = 31;
+		int result = 1;
+		result = prime * result + ((description == null) ? 0 : description.hashCode());
+		result = prime * result + ((id == null) ? 0 : id.hashCode());
+		result = prime * result + ((name == null) ? 0 : name.hashCode());
+		return result;
+	}
+
+	@Override
+	public boolean equals(Object obj) {
+		if (this == obj)
+			return true;
+		if (obj == null)
+			return false;
+		if (getClass() != obj.getClass())
+			return false;
+		RoleJPA other = (RoleJPA) obj;
+		if (description == null) {
+			if (other.description != null)
+				return false;
+		} else if (!description.equals(other.description))
+			return false;
+		if (id == null) {
+			if (other.id != null)
+				return false;
+		} else if (!id.equals(other.id))
+			return false;
+		if (name == null) {
+			if (other.name != null)
+				return false;
+		} else if (!name.equals(other.name))
+			return false;
+		return true;
+	}
+
+	@Override
+	public String toString() {
+		return "{\"id\":"+id+",\"name\":\"" + name + "\", \"description\":\"" + description + "\"}";
+	}	
+
+}
