+++ gvengine/gviam-db/src/main/java/it/greenvulcano/gvesb/iam/repository/hibernate/RepositoryHibernate.java
@@ -1,0 +1,18 @@
+package it.greenvulcano.gvesb.iam.repository.hibernate;
+
+import org.hibernate.Session;
+import org.hibernate.SessionFactory;
+
+abstract class RepositoryHibernate {
+	
+	private SessionFactory sessionFactory;
+			
+	protected void setSessionFactory(SessionFactory sessionFactory) {
+		this.sessionFactory = sessionFactory;
+	}
+
+	protected Session getSession() {
+		return sessionFactory.getCurrentSession();
+	}
+	
+}
