+++ gvengine/gviam/src/main/java/it/greenvulcano/gvesb/iam/domain/UserInfo.java
@@ -19,11 +19,6 @@
  *******************************************************************************/
 package it.greenvulcano.gvesb.iam.domain;
 
-import java.io.Serializable;
-
-import javax.persistence.Column;
-import javax.persistence.Embeddable;
-
 /**
  * 
  * Extended information about our {@link User}}
@@ -31,67 +26,13 @@ import javax.persistence.Embeddable;
  * @author GreenVulcano Technologies
  *
  */
-@Embeddable
-public class UserInfo implements Serializable {
+public abstract class UserInfo {
 	
-	private static final long serialVersionUID = 1L;
-		
-	@Column(length=256)
-	private String fullname;
-	
-	@Column(length=256)
-	private String email;
 
-	public String getFullname() {
-		return fullname;
-	}
+	public abstract String getFullname();
+	public abstract void setFullname(String fullname);
 
-	public void setFullname(String fullname) {
-		this.fullname = fullname;
-	}
-
-	public String getEmail() {
-		return email;
-	}
-
-	public void setEmail(String email) {
-		this.email = email;
-	}
-
-	@Override
-	public int hashCode() {
-		final int prime = 31;
-		int result = 1;
-		result = prime * result + ((email == null) ? 0 : email.hashCode());
-		result = prime * result + ((fullname == null) ? 0 : fullname.hashCode());
-		return result;
-	}
-
-	@Override
-	public boolean equals(Object obj) {
-		if (this == obj)
-			return true;
-		if (obj == null)
-			return false;
-		if (getClass() != obj.getClass())
-			return false;
-		UserInfo other = (UserInfo) obj;
-		if (email == null) {
-			if (other.email != null)
-				return false;
-		} else if (!email.equals(other.email))
-			return false;
-		if (fullname == null) {
-			if (other.fullname != null)
-				return false;
-		} else if (!fullname.equals(other.fullname))
-			return false;
-		return true;
-	}
-
-	@Override
-	public String toString() {
-		return "{\"fullname\":\"" + fullname + "\", \"email\":\"" + email + "\"}";
-	}	
+	public abstract String getEmail();
+	public abstract void setEmail(String email);
 
 }
