--- gvengine/gviam/src/main/java/it/greenvulcano/gvesb/iam/repository/RoleRepository.java
@@ -1,35 +1,0 @@
-/*******************************************************************************
- * Copyright (c) 2009, 2016 GreenVulcano ESB Open Source Project.
- * All rights reserved.
- *
- * This file is part of GreenVulcano ESB.
- *
- * GreenVulcano ESB is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Lesser General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * GreenVulcano ESB is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
- * GNU Lesser General Public License for more details.
- *
- * You should have received a copy of the GNU Lesser General Public License
- * along with GreenVulcano ESB. If not, see <http://www.gnu.org/licenses/>.
- *******************************************************************************/
-package it.greenvulcano.gvesb.iam.repository;
-
-import java.util.Optional;
-
-import it.greenvulcano.gvesb.iam.domain.Role;
-
-/**
- * Business interface to deal with {@link Role} entity
- * 
- */
-public interface RoleRepository extends Repository<Role, Integer> {
-	
-	Optional<Role> get(String rolename);
-			
-
-}
