+++ gvengine/pom.xml
@@ -194,6 +194,7 @@
     <module>gvconsole</module>
     <module>gvdte</module>
     <module>gviam</module>
+    <module>gviam-db</module>
     <module>gviamx</module>
     <module>gvmonitoring</module>
     <module>gvpolicy</module>
@@ -201,7 +202,7 @@
     <module>gvstatistics</module>
     <module>gvvcl</module>
     <module>gvvcl-internal</module>
-    <module>features</module>
+    <module>features</module>    
   </modules>
 
 </project>
